refactor: move eagle draft decode into draft worker
