+++ python/sgl_jax/srt/speculative/eagle_worker.py
@@ -1,4 +1,3 @@
-import functools
 import itertools
 import logging
 import time
@@ -6,11 +5,9 @@ import time
 import jax
 import jax.numpy as jnp
 import numpy as np
-from jax.sharding import NamedSharding
-from jax.sharding import PartitionSpec as P
 from tqdm import tqdm
 
-from sgl_jax.srt.layers.logits_processor import LogitsMetadata, LogitsProcessorOutput
+from sgl_jax.srt.layers.logits_processor import LogitsProcessorOutput
 from sgl_jax.srt.layers.sampler import get_token_ids_logprobs, get_top_logprobs
 from sgl_jax.srt.managers.schedule_batch import ModelWorkerBatch, ScheduleBatch
 from sgl_jax.srt.managers.scheduler import GenerationBatchResult
@@ -27,12 +24,10 @@ from sgl_jax.srt.speculative.eagle_util import (
     EagleDraftInput,
     EagleVerifyInput,
     EagleVerifyOutput,
-    build_tree_kernel_efficient,
-    build_tree_mask_for_draft_decode,
 )
 from sgl_jax.srt.speculative.spec_info import SpeculativeAlgorithm
+from sgl_jax.srt.speculative.spec_utils import topk_probs_from_logits
 from sgl_jax.srt.utils.common_utils import get_bool_env_var
-from sgl_jax.srt.utils.jax_utils import device_array
 
 logger = logging.getLogger(__name__)
 RETURN_ORIGINAL_LOGPROB = get_bool_env_var("RETURN_ORIGINAL_LOGPROB")
@@ -49,7 +44,6 @@ class EAGLEWorker(BaseSpecWorker):
         self.speculative_algorithm = SpeculativeAlgorithm.from_string(
             server_args.speculative_algorithm
         )
-        self.req_to_token_pool, self.token_to_kv_pool_allocator = target_worker.get_memory_pool()
         self._draft_worker = EagleDraftWorker(server_args, target_worker)
         self.precompile_bs_paddings = self._draft_worker.precompile_bs_paddings
         self.precompile_cache_loc_paddings = self._draft_worker.precompile_cache_loc_paddings
@@ -78,10 +72,6 @@ class EAGLEWorker(BaseSpecWorker):
     @property
     def model_runner(self):
         return self.draft_model_runner
-
-    @property
-    def hot_token_ids(self):
-        return self.draft_worker.hot_token_ids
 
     @property
     def max_req_len(self):
@@ -135,7 +125,7 @@ class EAGLEWorker(BaseSpecWorker):
 
         else:
             cur_allocate_lens = model_worker_batch.spec_info.allocate_lens
-            self.draft(model_worker_batch)
+            model_worker_batch.spec_info = self.draft_worker.draft(model_worker_batch)
 
             batch_output = self.verify(model_worker_batch, cur_allocate_lens)
 
@@ -159,228 +149,6 @@ class EAGLEWorker(BaseSpecWorker):
             model_worker_batch.bid,
             model_worker_batch.seq_lens,
         )
-
-    def copy_model_worker_batch_to_cpu(self, model_worker_batch: ModelWorkerBatch):
-        model_worker_batch.input_ids = np.array(
-            jax.device_get(model_worker_batch.input_ids), dtype=model_worker_batch.input_ids.dtype
-        )
-        model_worker_batch.seq_lens = np.array(
-            jax.device_get(model_worker_batch.seq_lens), dtype=model_worker_batch.seq_lens.dtype
-        )
-        model_worker_batch.out_cache_loc = np.array(
-            jax.device_get(model_worker_batch.out_cache_loc),
-            dtype=model_worker_batch.out_cache_loc.dtype,
-        )
-        model_worker_batch.positions = np.array(
-            jax.device_get(model_worker_batch.positions), dtype=model_worker_batch.positions.dtype
-        )
-        model_worker_batch.req_pool_indices = np.array(
-            jax.device_get(model_worker_batch.req_pool_indices),
-            dtype=model_worker_batch.req_pool_indices.dtype,
-        )
-        model_worker_batch.cache_loc = np.array(
-            jax.device_get(model_worker_batch.cache_loc), dtype=model_worker_batch.cache_loc.dtype
-        )
-        model_worker_batch.extend_prefix_lens = (
-            np.array(
-                jax.device_get(model_worker_batch.extend_prefix_lens),
-                dtype=model_worker_batch.extend_prefix_lens.dtype,
-            )
-            if model_worker_batch.extend_prefix_lens is not None
-            else None
-        )
-        model_worker_batch.extend_seq_lens = (
-            np.array(
-                jax.device_get(model_worker_batch.extend_seq_lens),
-                dtype=model_worker_batch.extend_seq_lens.dtype,
-            )
-            if model_worker_batch.extend_seq_lens is not None
-            else None
-        )
-
-    def capture_for_decode(
-        self, logits_output: LogitsProcessorOutput, draft_input: EagleDraftInput
-    ):
-        topk_p, topk_index = topk_probs_from_logits(logits_output.next_token_logits, self.topk)
-        draft_input.topk_p = topk_p
-        draft_input.topk_index = topk_index
-        draft_input.hidden_states = logits_output.hidden_states
-
-    def get_padding_bs(self, real_bs: int) -> int:
-        self.precompile_bs_paddings.sort()
-        select_bs_index = -1
-        bs_padding_size = 0
-        for i, size in enumerate(self.precompile_bs_paddings):
-            if size >= real_bs:
-                bs_padding_size = size - real_bs
-                select_bs_index = i
-                break
-        if select_bs_index < 0:
-            raise RuntimeError("did not get comperate padding bs, it should not happened")
-        return bs_padding_size, select_bs_index
-
-    def padding_for_decode(self, model_worker_batch: ModelWorkerBatch):
-        _, padding_bs_index = self.get_padding_bs(model_worker_batch.real_bs)
-        self.copy_model_worker_batch_to_cpu(model_worker_batch)
-        model_worker_batch.spec_info.prepare_for_draft_decode(
-            model_worker_batch, self.topk, self.speculative_num_steps
-        )
-        # get unpadded seq_lens
-        model_worker_batch.seq_lens = model_worker_batch.seq_lens
-        seq_lens_cpu = model_worker_batch.seq_lens
-        page_size = self.page_size
-        token_indices_with_all_reqs = self.req_to_token_pool.req_to_token[
-            model_worker_batch.req_pool_indices
-        ]
-        spec_info = model_worker_batch.spec_info
-        assert isinstance(spec_info, EagleDraftInput)
-        cache_loc_flat = np.array([], dtype=np.int32)
-        if len(seq_lens_cpu) > 0:
-            # Filter out empty sequences
-            valid_mask = seq_lens_cpu > 0
-            if np.any(valid_mask):
-                valid_indices = np.where(valid_mask)[0]
-                valid_allocate_lens = spec_info.allocate_lens[valid_mask]
-                # Calculate aligned lengths for all valid sequences at once
-                aligned_lengths = ((valid_allocate_lens + page_size - 1) // page_size) * page_size
-                total_aligned_length = np.sum(aligned_lengths)
-                # Pre-allocate the result array
-                cache_loc_flat = np.zeros(total_aligned_length, dtype=np.int32)
-                # Fill the array efficiently
-                offset = 0
-                for i, (seq_idx, allocate_len, aligned_len) in enumerate(
-                    zip(valid_indices, valid_allocate_lens, aligned_lengths)
-                ):
-                    # Copy the actual data
-                    cache_loc_flat[offset : offset + allocate_len] = token_indices_with_all_reqs[
-                        seq_idx, :allocate_len
-                    ]
-                    # Padding is already zero from initialization
-                    offset += aligned_len
-        total_cache_loc_size = self.precompile_cache_loc_paddings[padding_bs_index]
-        assert total_cache_loc_size >= len(cache_loc_flat)
-        cache_loc_cpu = np.empty(total_cache_loc_size, dtype=np.int32)
-        if len(cache_loc_flat) > 0:
-            cache_loc_cpu[: len(cache_loc_flat)] = cache_loc_flat
-        # Initialize padding area to ensure multiprocess consistency
-        if len(cache_loc_flat) < total_cache_loc_size:
-            cache_loc_cpu[len(cache_loc_flat) :] = 0
-
-        model_worker_batch.cache_loc = cache_loc_cpu
-        model_worker_batch.capture_hidden_mode = CaptureHiddenMode.LAST
-
-        # out_cache_loc = model_worker_batch.out_cache_loc
-        topk_index = spec_info.topk_index
-        if self.hot_token_ids is not None:
-            model_worker_batch.spec_info.topk_index = self.hot_token_ids[topk_index]
-        # if we need custom mask, we should create for all at once and update it within loop
-        # we should optimize build_tree_mask_for_draft_decode to a kernel
-        if self.topk > 1:
-            self.draft_model_runner.attn_backend.forward_metadata.custom_mask = (
-                build_tree_mask_for_draft_decode(
-                    model_worker_batch.seq_lens,
-                    topk=topk_index.shape[1],
-                    speculative_step_id=0,
-                    parents_list=None,
-                )
-            )
-        bs = self.precompile_bs_paddings[padding_bs_index]
-        if bs - model_worker_batch.spec_info.verified_id.shape[0] > 0:
-            model_worker_batch.spec_info.verified_id = np.pad(
-                model_worker_batch.spec_info.verified_id,
-                ((0, bs - model_worker_batch.spec_info.verified_id.shape[0]),),
-            )
-        if bs - model_worker_batch.spec_info.topk_p.shape[0] > 0:
-            model_worker_batch.spec_info.topk_p = np.pad(
-                model_worker_batch.spec_info.topk_p,
-                (
-                    (0, bs - model_worker_batch.spec_info.topk_p.shape[0]),
-                    (0, 0),
-                ),
-            )
-        if bs - model_worker_batch.seq_lens.shape[0] > 0:
-            model_worker_batch.seq_lens = np.pad(
-                model_worker_batch.seq_lens, ((0, bs - model_worker_batch.seq_lens.shape[0]),)
-            )
-            if model_worker_batch.spec_info.allocate_lens is not None:
-                model_worker_batch.spec_info.allocate_lens = np.pad(
-                    model_worker_batch.spec_info.allocate_lens,
-                    ((0, bs - model_worker_batch.spec_info.allocate_lens.shape[0]),),
-                )
-        if bs - model_worker_batch.spec_info.topk_index.shape[0] > 0:
-            model_worker_batch.spec_info.topk_index = np.pad(
-                model_worker_batch.spec_info.topk_index,
-                (
-                    (0, bs - model_worker_batch.spec_info.topk_index.shape[0]),
-                    (0, 0),
-                ),
-            )
-        if bs - model_worker_batch.spec_info.hidden_states.shape[0] > 0:
-            model_worker_batch.spec_info.hidden_states = np.pad(
-                model_worker_batch.spec_info.hidden_states,
-                (
-                    (0, bs - model_worker_batch.spec_info.hidden_states.shape[0]),
-                    (0, 0),
-                ),
-            )
-        # Forward multiple steps
-        model_worker_batch.speculative_eagle_topk = self.topk
-        model_worker_batch.speculative_num_steps = self.speculative_num_steps
-        model_worker_batch.speculative_num_draft_tokens = self.speculative_num_draft_tokens
-        model_worker_batch.input_ids = np.empty(bs * self.topk, np.int32)
-        model_worker_batch.positions = np.empty(bs * self.topk, np.int32)
-
-    def draft(self, model_worker_batch: ModelWorkerBatch):
-        self.padding_for_decode(model_worker_batch)
-        score_list, token_list, parents_list = self.draft_forward(model_worker_batch)
-        verified_seq_lens = model_worker_batch.seq_lens - 1
-        max_seq_len = int(np.max(verified_seq_lens)) if verified_seq_lens.size > 0 else 1
-        max_context_len = self._pick_context_len(max_seq_len)
-        (
-            tree_mask,
-            position,
-            retrive_index,
-            retrive_next_token,
-            retrive_next_sibling,
-            draft_tokens,
-        ) = build_tree_kernel_efficient(
-            model_worker_batch.spec_info.verified_id,
-            score_list,
-            token_list,
-            parents_list,
-            verified_seq_lens,
-            np.sum(verified_seq_lens),
-            self.topk,
-            self.speculative_num_draft_tokens,
-            max_context_len,
-            model_worker_batch.seq_lens.shape[0],
-            model_worker_batch.speculative_num_steps,
-            self.mesh,
-        )
-        model_worker_batch.spec_info = EagleVerifyInput(
-            draft_token=draft_tokens,
-            custom_mask=tree_mask,
-            positions=position,
-            retrive_index=retrive_index,
-            retrive_next_token=retrive_next_token,
-            retrive_next_sibling=retrive_next_sibling,
-            retrive_cum_len=None,
-            spec_steps=self.speculative_num_steps,
-            topk=self.topk,
-            draft_token_num=self.speculative_num_draft_tokens,
-            capture_hidden_mode=CaptureHiddenMode.LAST,
-            seq_lens_sum=model_worker_batch.seq_lens_sum,
-            seq_lens_cpu=model_worker_batch.seq_lens,
-        )
-        return model_worker_batch.spec_info
-
-    def _pick_context_len(self, max_seq_len: int) -> int:
-        max_seq_len = max(int(max_seq_len), 1)
-        if self.precompile_token_paddings:
-            for padding in self.precompile_token_paddings:
-                if padding >= max_seq_len:
-                    return padding
-        return 1 << (max_seq_len - 1).bit_length()
 
     def verify(self, model_worker_batch: ModelWorkerBatch, cur_allocate_lens: jax.Array):
         spec_info: EagleVerifyInput = model_worker_batch.spec_info
@@ -554,71 +322,6 @@ class EAGLEWorker(BaseSpecWorker):
         batch_output.allocate_lens = batch_output.allocate_lens[: model_worker_batch.real_bs]
         batch_output.accept_lens = batch_output.accept_lens[: model_worker_batch.real_bs]
 
-    def draft_forward(self, model_worker_batch: ModelWorkerBatch):
-        topk_p, topk_index, hidden_states = (
-            model_worker_batch.spec_info.topk_p,
-            model_worker_batch.spec_info.topk_index,
-            model_worker_batch.spec_info.hidden_states,
-        )
-        bs = model_worker_batch.seq_lens.shape[0]
-        step_min_1 = self.speculative_num_steps - 1
-        score_list: jax.Array = jnp.empty((bs, 1 + step_min_1 * self.topk, self.topk))
-        token_list: jax.Array = jnp.empty(
-            (bs, self.topk + step_min_1 * self.topk * self.topk), dtype=jnp.int32
-        )
-        parents_list: jax.Array = jnp.empty((bs, self.topk + 1 + step_min_1 * self.topk))
-        scores = None
-        positions_base = device_array(
-            np.repeat(model_worker_batch.seq_lens, self.topk),
-            sharding=(NamedSharding(self.model_runner.mesh, P())),
-        )
-        logits_metadata = None
-        metadata_per_step = self.draft_model_runner.attn_backend.get_eagle_multi_step_metadata(
-            model_worker_batch,
-        )
-        assert isinstance(metadata_per_step, list)
-        # we just use logits_metadata's forward mode and capture mode, it will not be modified within the loop
-        logits_metadata = LogitsMetadata.from_model_worker_batch(
-            model_worker_batch, self.draft_model_runner.mesh
-        )
-        forward_batch = ForwardBatch.init_new(model_worker_batch, self.draft_model_runner)
-        forward_batch.out_cache_loc = np.empty((1,))
-        forward_batch.cache_loc = np.empty((1,))
-        forward_batch.spec_info = EagleDraftInput()
-        forward_batch.spec_info.hidden_states = jnp.empty((bs * self.topk, hidden_states.shape[1]))
-        for i in range(self.speculative_num_steps):
-
-            input_ids, hidden_states, scores, tree_info = select_top_k_tokens(
-                i, topk_p, topk_index, hidden_states, scores, self.topk
-            )
-            # update_eagle_lists and update_forward_batch_info this two function will make accept rate very low if be jitted
-            # FIXME we should find it out why lead this ?
-            score_list, token_list, parents_list = update_eagle_lists(
-                i, score_list, token_list, parents_list, tree_info, self.topk
-            )
-            if i == self.speculative_num_steps - 1:
-                break
-
-            forward_batch = update_forward_batch_info(
-                forward_batch, i, input_ids, hidden_states, positions_base
-            )
-            self.draft_model_runner.attn_backend.forward_metadata = metadata_per_step[i]
-
-            # Run forward
-            forward_batch.bid = model_worker_batch.bid
-            logits_output, _, _ = self.draft_model_runner.forward(
-                forward_batch,
-                logits_metadata=logits_metadata,
-            )
-
-            topk_p, topk_index = topk_probs_from_logits(logits_output.next_token_logits, self.topk)
-
-            if self.hot_token_ids is not None:
-                topk_index = self.hot_token_ids[topk_index]
-            hidden_states = logits_output.hidden_states
-
-        return score_list, token_list, parents_list
-
     def run_spec_decode_precompile(self):
         self.precompile_spec_extend()
         self.precompile_spec_decode()
@@ -705,149 +408,3 @@ class EAGLEWorker(BaseSpecWorker):
 
         end_time = time.perf_counter()
         logger.info("[SPEC_DECODE] Precompile finished in %.0f secs", end_time - start_time)
-
-
-@functools.partial(jax.jit, static_argnames=["topk"])
-def topk_probs_from_logits(
-    logits: jax.Array, topk: int, axis: int = -1
-) -> tuple[jax.Array, jax.Array]:
-    """Return top-k probabilities without materializing the full softmax tensor."""
-    working_logits = jnp.moveaxis(logits, axis, -1) if axis != -1 else logits
-    topk_logits, topk_index = jax.lax.top_k(working_logits, topk)
-    logsumexp = jax.nn.logsumexp(working_logits, axis=-1, keepdims=True)
-    topk_probs = jnp.exp(topk_logits - logsumexp)
-
-    if axis != -1:
-        topk_probs = jnp.moveaxis(topk_probs, -1, axis)
-        topk_index = jnp.moveaxis(topk_index, -1, axis)
-
-    return topk_probs, topk_index
-
-
-def fast_topk(values, topk, axis=-1):
-    working_values = jnp.moveaxis(values, axis, -1) if axis != -1 else values
-    result_vals, result_indices = jax.lax.top_k(working_values, topk)
-
-    if axis != -1:
-        result_vals = jnp.moveaxis(result_vals, -1, axis)
-        result_indices = jnp.moveaxis(result_indices, -1, axis)
-
-    return result_vals, result_indices
-
-
-# FIXME(pc) this should be jitted or convert as np.ndarray
-# @functools.partial(jax.jit, static_argnames=["i", "topk"])
-def update_eagle_lists(
-    i: int,
-    score_list: jax.Array,
-    token_list: jax.Array,
-    parents_list: jax.Array,
-    tree_info: tuple[jax.Array, jax.Array, jax.Array],
-    topk: int,
-):
-    bs = score_list.shape[0]
-    scores_update, tokens_update, parents_update = tree_info
-    if i == 0:
-        score_list = score_list.at[:bs, :1, :].set(scores_update[:bs])
-        token_list = token_list.at[:bs, :topk].set(tokens_update[:bs])
-        parents_list = parents_list.at[:bs, : topk + 1].set(parents_update[:bs])
-    else:
-        score_start = 1 + (i - 1) * topk
-        token_start = topk + (i - 1) * topk * topk
-        parent_start = topk + 1 + (i - 1) * topk
-
-        score_list = score_list.at[:bs, score_start : score_start + topk, :].set(scores_update[:bs])
-        token_list = token_list.at[:bs, token_start : token_start + topk * topk].set(
-            tokens_update[:bs]
-        )
-        parents_list = parents_list.at[:bs, parent_start : parent_start + topk].set(
-            parents_update[:bs]
-        )
-    return score_list, token_list, parents_list
-
-
-# FIXME(pc) this should be jitted or convert as np.ndarray
-# @functools.partial(jax.jit, static_argnames=["i"])
-def update_forward_batch_info(
-    forward_batch: ForwardBatch,
-    i: int,
-    input_ids: jax.Array,
-    hidden_states: jax.Array,
-    positions_base: jax.Array,
-) -> ForwardBatch:
-    forward_batch.input_ids = input_ids
-    # FIXME(pc) hiddenstate will become NAN when forward path is very long, we still have no reason for this
-    forward_batch.spec_info.hidden_states = hidden_states
-    forward_batch.positions = positions_base + i
-    return forward_batch
-
-
-def select_top_k_tokens(
-    i: int,
-    topk_p: jax.Array,
-    topk_index: jax.Array,
-    hidden_states: jax.Array,
-    scores: jax.Array,
-    topk: int,
-) -> tuple[jax.Array, jax.Array, jax.Array, jax.Array]:
-    if i == 0:
-        return select_top_k_tokens_step_0(topk_p, topk_index, hidden_states, scores, topk)
-    else:
-        return select_top_k_tokens_step_greater_0(
-            jnp.asarray(i), topk_p, topk_index, hidden_states, scores, topk
-        )
-
-
-@functools.partial(jax.jit, static_argnames=["topk"])
-def select_top_k_tokens_step_0(
-    topk_p: jax.Array,
-    topk_index: jax.Array,
-    hidden_states: jax.Array,
-    scores: jax.Array,
-    topk: int,
-) -> tuple[jax.Array, jax.Array, jax.Array, jax.Array]:
-    # The first step after extend
-    input_ids = topk_index.flatten()
-    hidden_states = jnp.repeat(hidden_states, topk, axis=0)
-    scores = topk_p  # shape: (b, topk)
-    tree_info = (
-        jnp.expand_dims(topk_p, axis=1),  # shape: (b, 1, topk)
-        topk_index,  # shape: (b, topk)
-        jnp.tile(
-            jnp.expand_dims(jnp.arange(-1, topk, dtype=jnp.float32), axis=0),
-            (topk_p.shape[0], 1),
-        ),  # shape: (b, topk + 1)
-    )
-    return input_ids, hidden_states, scores, tree_info
-
-
-@functools.partial(jax.jit, static_argnames=["topk"])
-def select_top_k_tokens_step_greater_0(
-    i: jax.Array,
-    topk_p: jax.Array,
-    topk_index: jax.Array,
-    hidden_states: jax.Array,
-    scores: jax.Array,
-    topk: int,
-) -> tuple[jax.Array, jax.Array, jax.Array, jax.Array]:
-    # The later decode steps
-    expand_scores = jax.lax.mul(
-        jnp.expand_dims(scores, axis=2), topk_p.reshape(-1, topk, topk)
-    )  # (b, topk, 1) x (b, topk ,topk) -> (b, topk, topk)
-    topk_cs_p, topk_cs_index = fast_topk(
-        expand_scores.reshape(expand_scores.shape[0], -1), topk, axis=-1
-    )  # (b, topk)
-    scores = topk_cs_p  # shape: (b, topk)
-    topk_index = topk_index.reshape(-1, topk**2)
-    input_ids = jnp.take_along_axis(topk_index, topk_cs_index, axis=1).flatten()
-    if hidden_states.shape[0] > 0:
-        selected_input_index = topk_cs_index.flatten() // topk + jnp.repeat(
-            jnp.arange(0, hidden_states.shape[0], topk), topk
-        )
-        hidden_states = hidden_states[selected_input_index, :]
-    tree_info = (
-        expand_scores,  # shape: (b, topk, topk)
-        topk_index,  # shape: (b, topk * topk)
-        topk_cs_index + (topk**2 * (i - 1) + topk),  # shape: (b, topk)
-    )
-    return input_ids, hidden_states, scores, tree_info
