+++ python/sgl_jax/srt/speculative/spec_utils.py
@@ -1,0 +1,152 @@
+import functools
+
+import jax
+import jax.numpy as jnp
+
+from sgl_jax.srt.model_executor.forward_batch_info import ForwardBatch
+
+
+@functools.partial(jax.jit, static_argnames=["topk"])
+def topk_probs_from_logits(
+    logits: jax.Array, topk: int, axis: int = -1
+) -> tuple[jax.Array, jax.Array]:
+    """Return top-k probabilities without materializing the full softmax tensor."""
+    working_logits = jnp.moveaxis(logits, axis, -1) if axis != -1 else logits
+    topk_logits, topk_index = jax.lax.top_k(working_logits, topk)
+    logsumexp = jax.nn.logsumexp(working_logits, axis=-1, keepdims=True)
+    topk_probs = jnp.exp(topk_logits - logsumexp)
+
+    if axis != -1:
+        topk_probs = jnp.moveaxis(topk_probs, -1, axis)
+        topk_index = jnp.moveaxis(topk_index, -1, axis)
+
+    return topk_probs, topk_index
+
+
+def fast_topk(values, topk, axis=-1):
+    working_values = jnp.moveaxis(values, axis, -1) if axis != -1 else values
+    result_vals, result_indices = jax.lax.top_k(working_values, topk)
+
+    if axis != -1:
+        result_vals = jnp.moveaxis(result_vals, -1, axis)
+        result_indices = jnp.moveaxis(result_indices, -1, axis)
+
+    return result_vals, result_indices
+
+
+# FIXME(pc) this should be jitted or convert as np.ndarray
+# @functools.partial(jax.jit, static_argnames=["i", "topk"])
+def update_eagle_lists(
+    i: int,
+    score_list: jax.Array,
+    token_list: jax.Array,
+    parents_list: jax.Array,
+    tree_info: tuple[jax.Array, jax.Array, jax.Array],
+    topk: int,
+):
+    bs = score_list.shape[0]
+    scores_update, tokens_update, parents_update = tree_info
+    if i == 0:
+        score_list = score_list.at[:bs, :1, :].set(scores_update[:bs])
+        token_list = token_list.at[:bs, :topk].set(tokens_update[:bs])
+        parents_list = parents_list.at[:bs, : topk + 1].set(parents_update[:bs])
+    else:
+        score_start = 1 + (i - 1) * topk
+        token_start = topk + (i - 1) * topk * topk
+        parent_start = topk + 1 + (i - 1) * topk
+
+        score_list = score_list.at[:bs, score_start : score_start + topk, :].set(scores_update[:bs])
+        token_list = token_list.at[:bs, token_start : token_start + topk * topk].set(
+            tokens_update[:bs]
+        )
+        parents_list = parents_list.at[:bs, parent_start : parent_start + topk].set(
+            parents_update[:bs]
+        )
+    return score_list, token_list, parents_list
+
+
+# FIXME(pc) this should be jitted or convert as np.ndarray
+# @functools.partial(jax.jit, static_argnames=["i"])
+def update_forward_batch_info(
+    forward_batch: ForwardBatch,
+    i: int,
+    input_ids: jax.Array,
+    hidden_states: jax.Array,
+    positions_base: jax.Array,
+) -> ForwardBatch:
+    forward_batch.input_ids = input_ids
+    # FIXME(pc) hiddenstate will become NAN when forward path is very long, we still have no reason for this
+    forward_batch.spec_info.hidden_states = hidden_states
+    forward_batch.positions = positions_base + i
+    return forward_batch
+
+
+def select_top_k_tokens(
+    i: int,
+    topk_p: jax.Array,
+    topk_index: jax.Array,
+    hidden_states: jax.Array,
+    scores: jax.Array,
+    topk: int,
+) -> tuple[jax.Array, jax.Array, jax.Array, jax.Array]:
+    if i == 0:
+        return select_top_k_tokens_step_0(topk_p, topk_index, hidden_states, scores, topk)
+    else:
+        return select_top_k_tokens_step_greater_0(
+            jnp.asarray(i), topk_p, topk_index, hidden_states, scores, topk
+        )
+
+
+@functools.partial(jax.jit, static_argnames=["topk"])
+def select_top_k_tokens_step_0(
+    topk_p: jax.Array,
+    topk_index: jax.Array,
+    hidden_states: jax.Array,
+    scores: jax.Array,
+    topk: int,
+) -> tuple[jax.Array, jax.Array, jax.Array, jax.Array]:
+    # The first step after extend
+    input_ids = topk_index.flatten()
+    hidden_states = jnp.repeat(hidden_states, topk, axis=0)
+    scores = topk_p  # shape: (b, topk)
+    tree_info = (
+        jnp.expand_dims(topk_p, axis=1),  # shape: (b, 1, topk)
+        topk_index,  # shape: (b, topk)
+        jnp.tile(
+            jnp.expand_dims(jnp.arange(-1, topk, dtype=jnp.float32), axis=0),
+            (topk_p.shape[0], 1),
+        ),  # shape: (b, topk + 1)
+    )
+    return input_ids, hidden_states, scores, tree_info
+
+
+@functools.partial(jax.jit, static_argnames=["topk"])
+def select_top_k_tokens_step_greater_0(
+    i: jax.Array,
+    topk_p: jax.Array,
+    topk_index: jax.Array,
+    hidden_states: jax.Array,
+    scores: jax.Array,
+    topk: int,
+) -> tuple[jax.Array, jax.Array, jax.Array, jax.Array]:
+    # The later decode steps
+    expand_scores = jax.lax.mul(
+        jnp.expand_dims(scores, axis=2), topk_p.reshape(-1, topk, topk)
+    )  # (b, topk, 1) x (b, topk ,topk) -> (b, topk, topk)
+    topk_cs_p, topk_cs_index = fast_topk(
+        expand_scores.reshape(expand_scores.shape[0], -1), topk, axis=-1
+    )  # (b, topk)
+    scores = topk_cs_p  # shape: (b, topk)
+    topk_index = topk_index.reshape(-1, topk**2)
+    input_ids = jnp.take_along_axis(topk_index, topk_cs_index, axis=1).flatten()
+    if hidden_states.shape[0] > 0:
+        selected_input_index = topk_cs_index.flatten() // topk + jnp.repeat(
+            jnp.arange(0, hidden_states.shape[0], topk), topk
+        )
+        hidden_states = hidden_states[selected_input_index, :]
+    tree_info = (
+        expand_scores,  # shape: (b, topk, topk)
+        topk_index,  # shape: (b, topk * topk)
+        topk_cs_index + (topk**2 * (i - 1) + topk),  # shape: (b, topk)
+    )
+    return input_ids, hidden_states, scores, tree_info
